Split Android sync local store contract tests
